+++ backend/canisters/user_index/impl/src/lifecycle/post_upgrade.rs
@@ -1,12 +1,15 @@
 use crate::lifecycle::{init_env, init_state};
 use crate::memory::get_upgrades_memory;
-use crate::{mutate_state, Data};
+use crate::updates::unsuspend_user::unsuspend_user_impl;
+use crate::Data;
+use candid::Principal;
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use ic_cdk_macros::post_upgrade;
-use local_user_index_canister::Event;
 use stable_memory::get_reader;
+use std::time::Duration;
 use tracing::info;
+use types::UserId;
 use user_index_canister::post_upgrade::Args;
 use utils::cycles::init_cycles_dispenser_client;
 
@@ -24,22 +27,23 @@ fn post_upgrade(args: Args) {
     init_cycles_dispenser_client(data.cycles_dispenser_canister_id, data.test_mode);
     init_state(env, data, args.wasm_version);
 
-    mutate_state(|state| {
-        let now = state.env.now();
-        for user in state.data.users.iter() {
-            if let Some(expires_at) = user.diamond_membership_details.expires_at() {
-                if expires_at > now {
-                    for local_user_index_canister_id in state.data.local_index_map.canisters() {
-                        state.data.user_index_event_sync_queue.push(
-                            *local_user_index_canister_id,
-                            Event::DiamondMembershipExpiryDate(user.user_id, expires_at),
-                        );
-                    }
-                }
-            }
-        }
-        crate::jobs::sync_events_to_local_user_index_canisters::start_job_if_required(state);
-    });
-
     info!(version = %args.wasm_version, "Post-upgrade complete");
+
+    let users_to_unsuspend: Vec<UserId> = [
+        "qyzts-uiaaa-aaaar-aw4ca-cai",
+        "zjayq-ryaaa-aaaar-axhrq-cai",
+        "o6334-oyaaa-aaaar-awgxa-cai",
+        "s4i57-nqaaa-aaaar-abzcq-cai",
+        "vecg6-eqaaa-aaaar-a23rq-cai",
+        "ne6gg-dqaaa-aaaaf-agzwa-cai",
+    ]
+    .iter()
+    .map(|str| Principal::from_text(str).unwrap().into())
+    .collect();
+
+    ic_cdk_timers::set_timer(Duration::ZERO, || {
+        ic_cdk::spawn(async move {
+            futures::future::join_all(users_to_unsuspend.into_iter().map(unsuspend_user_impl)).await;
+        })
+    });
 }