+++ frontend/openchat-client/src/openchat.ts
@@ -993,8 +993,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp === "success";
             })
-            .catch((err) => {
-                this._logger.error("Error toggling mute notifications", err);
+            .catch(() => {
                 localChatSummaryUpdates.markUpdated(chatId, { notificationsMuted: undefined });
                 return false;
             });
@@ -1006,8 +1005,7 @@ export class OpenChat extends OpenChatAgentWorker {
             .then((resp) => {
                 return resp === "success";
             })
-            .catch((err) => {
-                this._logger.error("Error archiving chat", err);
+            .catch(() => {
                 localChatSummaryUpdates.markUpdated(chatId, { archived: undefined });
                 return false;
             });
@@ -1017,8 +1015,7 @@ export class OpenChat extends OpenChatAgentWorker {
         localChatSummaryUpdates.markUpdated(chatId, { archived: false });
         return this.sendRequest({ kind: "unarchiveChat", chatId })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Error un-archiving chat", err);
+            .catch(() => {
                 localChatSummaryUpdates.markUpdated(chatId, { archived: undefined });
                 return false;
             });
@@ -1073,8 +1070,7 @@ export class OpenChat extends OpenChatAgentWorker {
             favourite: scope === "favourite",
         })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Error pinning chat", err);
+            .catch(() => {
                 this.unpinLocally(chatId, scope);
                 return false;
             });
@@ -1089,8 +1085,7 @@ export class OpenChat extends OpenChatAgentWorker {
             favourite: scope === "favourite",
         })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Error unpinning chat", err);
+            .catch(() => {
                 this.pinLocally(chatId, scope);
                 return false;
             });
@@ -1102,8 +1097,7 @@ export class OpenChat extends OpenChatAgentWorker {
             .then((resp) => {
                 return resp === "success";
             })
-            .catch((err) => {
-                this._logger.error("Error blocking user", err);
+            .catch(() => {
                 blockedUsers.delete(userId);
                 return false;
             });
@@ -1115,8 +1109,7 @@ export class OpenChat extends OpenChatAgentWorker {
             .then((resp) => {
                 return resp === "success";
             })
-            .catch((err) => {
-                this._logger.error("Error unblocking user", err);
+            .catch(() => {
                 blockedUsers.add(userId);
                 return false;
             });
@@ -1134,10 +1127,7 @@ export class OpenChat extends OpenChatAgentWorker {
 
         return this.sendRequest({ kind: "setUserAvatar", data })
             .then((_resp) => true)
-            .catch((err) => {
-                this._logger.error("Failed to update user's avatar", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     deleteGroup(chatId: MultiUserChatIdentifier): Promise<boolean> {
@@ -1151,10 +1141,7 @@ export class OpenChat extends OpenChatAgentWorker {
                     return false;
                 }
             })
-            .catch((err) => {
-                this._logger.error("Unable to delete group", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     deleteDirectChat(userId: string, blockUser: boolean): Promise<boolean> {
@@ -1191,10 +1178,7 @@ export class OpenChat extends OpenChatAgentWorker {
                     }
                 }
             })
-            .catch((err) => {
-                this._logger.error("Unable to leave group", err);
-                return "failure";
-            });
+            .catch(() => "failure");
     }
 
     private addCommunityLocally(community: CommunitySummary): void {
@@ -1332,20 +1316,14 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp;
             })
-            .catch((err) => {
-                this._logger.error("Unable to join group", err);
-                return "failure";
-            });
+            .catch(() => "failure");
     }
 
     setCommunityIndexes(indexes: Record<string, number>): Promise<boolean> {
         Object.entries(indexes).forEach(([k, v]) =>
             localCommunitySummaryUpdates.updateIndex({ kind: "community", communityId: k }, v),
         );
-        return this.sendRequest({ kind: "setCommunityIndexes", indexes }).catch((err) => {
-            this._logger.error("Failed to set community indexes: ", err);
-            return false;
-        });
+        return this.sendRequest({ kind: "setCommunityIndexes", indexes }).catch(() => false);
     }
 
     setMemberDisplayName(
@@ -1766,10 +1744,7 @@ export class OpenChat extends OpenChatAgentWorker {
             threadRootMessageIndex,
         })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Poll vote failed: ", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     deleteMessage(
@@ -1822,9 +1797,8 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return success;
             })
-            .catch((err) => {
+            .catch(() => {
                 _undelete();
-                this._logger.error("Delete message failed: ", err);
                 return false;
             });
     }
@@ -1856,10 +1830,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return success;
             })
-            .catch((err) => {
-                this._logger.error("Undelete message failed: ", err);
-                return false;
-            })
+            .catch(() => false)
             .finally(() => {
                 undeletingMessagesStore.delete(msg.messageId);
             });
@@ -1898,10 +1869,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return success;
             })
-            .catch((err) => {
-                this._logger.error("Get deleted message failed: ", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     revealBlockedMessage(messageId: bigint) {
@@ -2215,8 +2183,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return true;
             })
-            .catch((err) => {
-                this._logger.error("Error blocking community user", err);
+            .catch(() => {
                 this.unblockCommunityUserLocally(id, userId, true);
                 return false;
             });
@@ -2232,8 +2199,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return true;
             })
-            .catch((err) => {
-                this._logger.error("Error blocking community user", err);
+            .catch(() => {
                 this.blockCommunityUserLocally(id, userId);
                 return false;
             });
@@ -2250,8 +2216,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return true;
             })
-            .catch((err) => {
-                this._logger.error("Error blocking user", err);
+            .catch(() => {
                 this.unblockUserLocally(chatId, userId, true);
                 return false;
             });
@@ -2267,8 +2232,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return true;
             })
-            .catch((err) => {
-                this._logger.error("Error blocking user", err);
+            .catch(() => {
                 this.blockUserLocally(chatId, userId);
                 return false;
             });
@@ -2956,8 +2920,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return true;
             })
-            .catch((err) => {
-                this._logger.error("Unpin message failed: ", err);
+            .catch(() => {
                 this.addPinnedMessage(chatId, messageIndex);
                 return false;
             });
@@ -2980,8 +2943,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return true;
             })
-            .catch((err) => {
-                this._logger.error("Pin message failed: ", err);
+            .catch(() => {
                 this.removePinnedMessage(chatId, messageIndex);
                 return false;
             });
@@ -3578,8 +3540,7 @@ export class OpenChat extends OpenChatAgentWorker {
                     }
                     return true;
                 })
-                .catch((err) => {
-                    this._logger.error("Exception sending message", err);
+                .catch(() => {
                     localMessageUpdates.revertEditedContent(msg.messageId);
                     return false;
                 });
@@ -3617,8 +3578,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return true;
             })
-            .catch((err) => {
-                this._logger.error("Exception sending message", err);
+            .catch(() => {
                 localMessageUpdates.revertEditedContent(msg.messageId);
                 return false;
             });
@@ -4061,8 +4021,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp;
             })
-            .catch((err) => {
-                this._logger.error("Error uninviting users", err);
+            .catch(() => {
                 this.uninviteUsersLocally(chatId, userIds);
                 return "failure";
             });
@@ -4091,8 +4050,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp;
             })
-            .catch((err) => {
-                this._logger.error("Error inviting users to community", err);
+            .catch(() => {
                 this.uninviteUsersToCommunityLocally(id, userIds);
                 return "failure";
             });
@@ -4133,10 +4091,7 @@ export class OpenChat extends OpenChatAgentWorker {
             .then((resp) => {
                 return resp === "success";
             })
-            .catch((err) => {
-                this._logger.error("Error trying to change role: ", err);
-                return false;
-            })
+            .catch(() => false)
             .then((success) => {
                 if (!success) {
                     // Revert the local store
@@ -4169,10 +4124,7 @@ export class OpenChat extends OpenChatAgentWorker {
             .then((resp) => {
                 return resp === "success";
             })
-            .catch((err) => {
-                this._logger.error("Error trying to change role: ", err);
-                return false;
-            })
+            .catch(() => false)
             .then((success) => {
                 if (!success) {
                     // Revert the local store
@@ -4583,10 +4535,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return false;
             })
-            .catch((err) => {
-                this._logger.error("Unable to freeze group", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     unfreezeGroup(chatId: GroupChatIdentifier): Promise<boolean> {
@@ -4598,44 +4547,32 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return false;
             })
-            .catch((err) => {
-                this._logger.error("Unable to unfreeze group", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     deleteFrozenGroup(chatId: GroupChatIdentifier): Promise<boolean> {
         return this.sendRequest({ kind: "deleteFrozenGroup", chatId })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Unable to unfreeze group", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     addHotGroupExclusion(chatId: GroupChatIdentifier): Promise<boolean> {
         return this.sendRequest({ kind: "addHotGroupExclusion", chatId })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Unable to add hot group exclusion", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     removeHotGroupExclusion(chatId: GroupChatIdentifier): Promise<boolean> {
         return this.sendRequest({ kind: "removeHotGroupExclusion", chatId })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Unable to remove hot group exclusion", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     addMessageFilter(regex: string): Promise<boolean> {
         try {
             new RegExp(regex);
         } catch (e) {
-            this._logger.error("Unable to add message filter - invalid regex", regex);
+            console.error("Unable to add message filter - invalid regex", regex);
             return Promise.resolve(false);
         }
 
@@ -4649,62 +4586,41 @@ export class OpenChat extends OpenChatAgentWorker {
     suspendUser(userId: string, reason: string): Promise<boolean> {
         return this.sendRequest({ kind: "suspendUser", userId, reason })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Unable to suspend user", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     unsuspendUser(userId: string): Promise<boolean> {
         return this.sendRequest({ kind: "unsuspendUser", userId })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Unable to un-suspend user", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     setCommunityModerationFlags(communityId: string, flags: number): Promise<boolean> {
         return this.sendRequest({ kind: "setCommunityModerationFlags", communityId, flags })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Unable to set community moderation flags", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     setGroupUpgradeConcurrency(value: number): Promise<boolean> {
         return this.sendRequest({ kind: "setGroupUpgradeConcurrency", value })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Unable to set group upgrade concurrency", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     setCommunityUpgradeConcurrency(value: number): Promise<boolean> {
         return this.sendRequest({ kind: "setCommunityUpgradeConcurrency", value })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Unable to set group community concurrency", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     setUserUpgradeConcurrency(value: number): Promise<boolean> {
         return this.sendRequest({ kind: "setUserUpgradeConcurrency", value })
             .then((resp) => resp === "success")
-            .catch((err) => {
-                this._logger.error("Unable to set user upgrade concurrency", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     setDiamondMembershipFees(fees: DiamondMembershipFees[]): Promise<boolean> {
-        return this.sendRequest({ kind: "setDiamondMembershipFees", fees }).catch((err) => {
-            this._logger.error("Unable to set diamond membership fees", err);
-            return false;
-        });
+        return this.sendRequest({ kind: "setDiamondMembershipFees", fees }).catch(() => false);
     }
 
     stakeNeuronForSubmittingProposals(
@@ -4717,10 +4633,7 @@ export class OpenChat extends OpenChatAgentWorker {
             stake,
         })
             .then((resp) => resp.kind === "success")
-            .catch((err) => {
-                this._logger.error("Failed to stake neuron for submitting proposals", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     private onChatFrozen(
@@ -4966,7 +4879,7 @@ export class OpenChat extends OpenChatAgentWorker {
 
         const updateRegistryTask = initialLoad ? this.updateRegistry() : undefined;
 
-        return new Promise<void>((resolve) => {
+        return new Promise<void>((resolve, reject) => {
             this.sendStreamRequest({
                 kind: "getUpdates",
                 initialLoad,
@@ -4980,8 +4893,7 @@ export class OpenChat extends OpenChatAgentWorker {
                     chatsLoading.set(!this._liveState.chatsInitialised);
                 })
                 .catch((err) => {
-                    this.config.logger.error("Error loading chats: ", err);
-                    throw err;
+                    reject(err);
                 })
                 .finally(() => {
                     resolve();
@@ -5069,10 +4981,7 @@ export class OpenChat extends OpenChatAgentWorker {
                     return true;
                 }
             })
-            .catch((err) => {
-                this._logger.error("Claiming prize failed", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     acceptP2PSwap(
@@ -5099,7 +5008,6 @@ export class OpenChat extends OpenChatAgentWorker {
             })
             .catch((err) => {
                 localMessageUpdates.setP2PSwapStatus(messageId, { kind: "p2p_swap_open" });
-                this._logger.error("Accepting p2p swap failed", err);
                 return { kind: "internal_error", text: err.toString() };
             });
     }
@@ -5127,7 +5035,6 @@ export class OpenChat extends OpenChatAgentWorker {
             })
             .catch((err) => {
                 localMessageUpdates.setP2PSwapStatus(messageId, { kind: "p2p_swap_open" });
-                this._logger.error("Cancelling p2p swap failed", err);
                 return { kind: "internal_error", text: err.toString() };
             });
     }
@@ -5218,10 +5125,7 @@ export class OpenChat extends OpenChatAgentWorker {
                     return true;
                 }
             })
-            .catch((err) => {
-                this._logger.error("Paying for diamond membership failed", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     setMessageReminder(
@@ -5242,10 +5146,7 @@ export class OpenChat extends OpenChatAgentWorker {
             .then((res) => {
                 return res === "success";
             })
-            .catch((err) => {
-                this._logger.error("Unable to set message reminder", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     cancelMessageReminder(
@@ -5256,9 +5157,8 @@ export class OpenChat extends OpenChatAgentWorker {
         return this.sendRequest({
             kind: "cancelMessageReminder",
             reminderId: content.reminderId,
-        }).catch((err) => {
+        }).catch(() => {
             localMessageUpdates.revertCancelled(messageId);
-            this._logger.error("Unable to cancel message reminder", err);
             return false;
         });
     }
@@ -5275,10 +5175,7 @@ export class OpenChat extends OpenChatAgentWorker {
             threadRootMessageIndex,
             messageId,
             deleteMessage,
-        }).catch((err) => {
-            this._logger.error("Unable to report message", err);
-            return false;
-        });
+        }).catch(() => false);
     }
 
     declineInvitation(chatId: MultiUserChatIdentifier): Promise<boolean> {
@@ -5286,10 +5183,7 @@ export class OpenChat extends OpenChatAgentWorker {
             .then((res) => {
                 return res === "success";
             })
-            .catch((err) => {
-                this._logger.error("Failed to decline invitation", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     updateMarketMakerConfig(
@@ -5324,8 +5218,7 @@ export class OpenChat extends OpenChatAgentWorker {
             flags,
         })
             .then((resp) => (resp === "success" ? flags : previousValue))
-            .catch((err) => {
-                this._logger.error("Error setting moderation flags", err);
+            .catch(() => {
                 this.user.update((user) => ({
                     ...user,
                     moderationFlagsEnabled: previousValue,
@@ -5435,14 +5328,16 @@ export class OpenChat extends OpenChatAgentWorker {
         });
     }
 
-    private async updateExchangeRates(): Promise<void> {
-        const exchangeRates = await this.sendRequest({ kind: "exchangeRates" });
+    private updateExchangeRates(): Promise<void> {
+        return this.sendRequest({ kind: "exchangeRates" })
+            .then((exchangeRates) => {
+                // Handle couple of special cases
+                exchangeRates["dkp"] = exchangeRates["sns1"];
+                exchangeRates["icp"] = { ...exchangeRates["icp"], toICP: 1 };
 
-        // Handle couple of special cases
-        exchangeRates["dkp"] = exchangeRates["sns1"];
-        exchangeRates["icp"] = { ...exchangeRates["icp"], toICP: 1 };
-
-        exchangeRatesLookupStore.set(exchangeRates);
+                exchangeRatesLookupStore.set(exchangeRates);
+            })
+            .catch(() => undefined);
     }
 
     private async refreshBalancesInSeries() {
@@ -5567,10 +5462,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 this._logger.error("Failed to submit proposal", resp);
                 return false;
             })
-            .catch((err) => {
-                this._logger.error("Unable to submit proposal", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     swappableTokens(): Promise<Set<string>> {
@@ -5810,10 +5702,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return undefined;
             })
-            .catch((err) => {
-                this._logger.error("Unable to import group to community", err);
-                return undefined;
-            });
+            .catch(() => undefined);
     }
 
     async joinCommunity(
@@ -5855,10 +5744,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return "success";
             })
-            .catch((err) => {
-                this._logger.error("Unable to join community", err);
-                return "failure";
-            });
+            .catch(() => "failure");
     }
 
     deleteCommunity(id: CommunityIdentifier): Promise<boolean> {
@@ -5874,10 +5760,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp === "success";
             })
-            .catch((err) => {
-                this._logger.error("Error deleting community", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     leaveCommunity(id: CommunityIdentifier): Promise<boolean> {
@@ -5893,10 +5776,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp === "success";
             })
-            .catch((err) => {
-                this._logger.error("Error leaving community", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     createCommunity(
@@ -5910,10 +5790,9 @@ export class OpenChat extends OpenChatAgentWorker {
             rules,
             defaultChannels,
             defaultChannelRules: defaultChatRules("channel"),
-        }).catch((err) => {
-            this._logger.error("Error creating community", err);
-            return { kind: "failure" };
-        });
+        }).catch(() => ({
+            kind: "failure",
+        }));
     }
 
     private addToFavouritesLocally(chatId: ChatIdentifier): void {
@@ -5939,9 +5818,8 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp === "success";
             })
-            .catch((err) => {
+            .catch(() => {
                 this.removeFromFavouritesLocally(chatId);
-                this._logger.error("Error adding chat to favourites", err);
                 return false;
             });
     }
@@ -5955,9 +5833,8 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp === "success";
             })
-            .catch((err) => {
+            .catch(() => {
                 this.addToFavouritesLocally(chatId);
-                this._logger.error("Error removing chat from favourites", err);
                 return false;
             });
     }
@@ -6004,10 +5881,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return false;
             })
-            .catch((err) => {
-                this._logger.error("Error creating community", err);
-                return false;
-            });
+            .catch(() => false);
     }
 
     convertGroupToCommunity(
@@ -6021,10 +5895,7 @@ export class OpenChat extends OpenChatAgentWorker {
             rules,
         })
             .then((resp) => (resp.kind === "success" ? resp.id : undefined))
-            .catch((err) => {
-                this._logger.error("Error converting group to community", err);
-                return undefined;
-            });
+            .catch(() => undefined);
     }
 
     private deleteUserGroupLocally(id: CommunityIdentifier, userGroup: UserGroupDetails) {
@@ -6054,9 +5925,8 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp.kind === "success";
             })
-            .catch((err) => {
+            .catch(() => {
                 this.undeleteUserGroupLocally(id, userGroup);
-                this._logger.error("Error deleting community user group", err);
                 return false;
             });
     }
@@ -6080,10 +5950,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp;
             })
-            .catch((err) => {
-                this._logger.error("Error creating community user group", err);
-                return CommonResponses.failure();
-            });
+            .catch(() => CommonResponses.failure());
     }
 
     getCommunityForChannel(id: ChannelIdentifier): CommunitySummary | undefined {
@@ -6115,10 +5982,7 @@ export class OpenChat extends OpenChatAgentWorker {
                 }
                 return resp;
             })
-            .catch((err) => {
-                this._logger.error("Error updating community user group", err);
-                return CommonResponses.failure();
-            });
+            .catch(() => CommonResponses.failure());
     }
 
     setChatListScope(scope: ChatListScope): void {